Готовое решение. исправил все критические ошибки и замечания, которые были в Git
