+++ MovieQuiz.xcodeproj/xcuserdata/sergey.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "1C05A93E-3C00-4BF3-AE2D-2523C66710DC"
+   type = "1"
+   version = "2.0">
+</Bucket>
