+++ MovieQuiz/Models/File 2.swift
@@ -1,0 +1,9 @@
+//
+//  File 2.swift
+//  MovieQuiz
+//
+//  Created by Sergey on 18.03.2025.
+//
+
+
+import UIKit
