+++ MovieQuiz/Models/File.swift
@@ -1,0 +1,9 @@
+//
+//  File.swift
+//  MovieQuiz
+//
+//  Created by Sergey on 18.03.2025.
+//
+
+
+import UIKit
