+++ MovieQuiz/Models/QuizResultsModel.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+struct QuizResultsModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
