+++ MovieQuiz/Models/QuizStepModel.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+struct QuizStepModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
