+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -1,89 +1,49 @@
 import UIKit
 
 final class MovieQuizViewController: UIViewController {
-    // MARK: - Lifecycle
-   
+    // MARK: - Properties
+  
+    @IBOutlet private var imageView: UIImageView!
+    @IBOutlet private var textLabel: UILabel!
+    @IBOutlet private var counterLabel: UILabel!
+    
     private var currentQuestionIndex = 0
     private var correctAnswers = 0
+    private let questions: [QuizQuestion] = QuizQuestionMock.questions
     
-    private let questions: [QuizQuestion] = [
-            QuizQuestion(
-                image: "The Godfather",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "The Dark Knight",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "Kill Bill",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "The Avengers",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "Deadpool",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "The Green Knight",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "Old",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: false),
-            QuizQuestion(
-                image: "The Ice Age Adventures of Buck Wild",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: false),
-            QuizQuestion(
-                image: "Tesla",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: false),
-            QuizQuestion(
-                image: "Vivarium",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: false)
-        ]
-    
+    // MARK: - Lifecycle
     
     override func viewDidLoad() {
         super.viewDidLoad()
         viewNext()
     }
     
+    // MARK: - Setup Methods
+    
     private func viewNext () {
         let currentQuestion = questions[currentQuestionIndex]
         let viewModel = convert(model: currentQuestion)
         show(quiz: viewModel)
-        imageView.layer.borderColor = UIColor.ypWhite.cgColor
-        
+        imageView.layer.borderColor = UIColor.clear.cgColor
     }
     
-    private func convert(model: QuizQuestion) -> QuizStepViewModel {
-        let questionStep = QuizStepViewModel(
+    private func convert(model: QuizQuestion) -> QuizStepModel {
+        let questionStep = QuizStepModel(
             image: UIImage(named: model.image) ?? UIImage(),
             question: model.text,
             questionNumber: "\(currentQuestionIndex + 1)/\(questions.count)")
         return questionStep
-        
     }
   
-    
-    private func show(quiz step: QuizStepViewModel) {
+    private func show(quiz step: QuizStepModel) {
         imageView.image = step.image
         textLabel.text = step.question
         counterLabel.text = step.questionNumber
-  
     }
    
     private func showAnswerResult(isCorrect: Bool) {
         imageView.layer.masksToBounds = true
         imageView.layer.borderWidth = 8
-        imageView.layer.borderColor = UIColor.white.cgColor
         imageView.layer.cornerRadius = 20
       
         if isCorrect {
@@ -95,7 +55,6 @@ final class MovieQuizViewController: UIViewController {
         
         DispatchQueue.main.asyncAfter(deadline: .now() + 1) {
             self.showNextQuestionOrResults()
-            
         }
     }
     
@@ -110,7 +69,7 @@ final class MovieQuizViewController: UIViewController {
     private func showNextQuestionOrResults() {
         if currentQuestionIndex == questions.count - 1 {
             let resultFinal = "Ваш результат: \(correctAnswers) / \(questions.count)"
-            let final = QuizResultsViewModel(title: "Этот раунд окончен!" ,
+            let final = QuizResultsModel(title: "Этот раунд окончен!" ,
                                              text: resultFinal,
                                              buttonText:"Сыграть ещё раз" )
             show(quiz: final)
@@ -120,8 +79,7 @@ final class MovieQuizViewController: UIViewController {
             }
     }
   
-
-    private func show(quiz result: QuizResultsViewModel) {
+    private func show(quiz result: QuizResultsModel) {
         let alert = UIAlertController(
             title: result.title,
             message: result.text,
@@ -137,41 +95,19 @@ final class MovieQuizViewController: UIViewController {
         self.present(alert, animated: true, completion: nil)
     }
     
+    // MARK: - Actions
     
     @IBAction private func yesButtonClicked(_ sender: UIButton) {
         let currentQuestion = questions[currentQuestionIndex]
         let result = true
     showAnswerResult(isCorrect: result == currentQuestion.correctAnswer)
         stopButtonTapped(sender: sender)
-       
     }
+    
     @IBAction private func noButtonClicked(_ sender: UIButton) {
         let currentQuestion = questions[currentQuestionIndex]
         let result = false
         showAnswerResult(isCorrect: result == currentQuestion.correctAnswer)
         stopButtonTapped(sender: sender)
-    }
-    
-    @IBOutlet private var imageView: UIImageView!
-    @IBOutlet private var textLabel: UILabel!
-    @IBOutlet private var counterLabel: UILabel!
-    
-    
-    struct QuizStepViewModel {
-        let image: UIImage
-        let question: String
-        let questionNumber: String
-    }
-    
-    struct QuizResultsViewModel {
-        let title: String
-        let text: String
-        let buttonText: String
-    }
-    
-    struct QuizQuestion {
-        let image: String
-        let text: String
-        let correctAnswer: Bool
     }
 }
